shotgame/requirements.txt
Django==4.2.18
djangorestframework==3.16.1
groq==0.34.1
python-dotenv==1.1.0
PyMySQL==1.1.1
requests==2.32.5
gunicorn==23.0.0
whitenoise==6.11.0
certifi